Javascript fixes
Updated Gulp file to include a new task to minify the datatables
Javascript file and place it in the correct folder. The Gulp task is
also modified as it now has two different files to create. A broken link
to the file is also fixed in the account section footer, and the
minified file has been added.